Refactored 2fa send email and identity to cater for passwordless
