+++ src/Api/Models/Request/TwoFactorRequestModels.cs
@@ -205,7 +205,6 @@ public class TwoFactorEmailRequestModel : SecretVerificationRequestModel
     public string DeviceIdentifier { get; set; }
 
     public Guid AuthRequestId { get; set; }
-    public string AuthRequestAccessCode { get; set; }
 
     public User ToUser(User extistingUser)
     {
