+++ src/Core/LoginFeatures/PasswordlessLogin/VerifyAuthRequest.cs
@@ -1,10 +1,7 @@
-using Bit.Core.Entities;
-using Bit.Core.Enums;
-using Bit.Core.OrganizationFeatures.OrganizationApiKeys.Interfaces;
+using Bit.Core.LoginFeatures.PasswordlessLogin.Interfaces;
 using Bit.Core.Repositories;
-using Bit.Core.Utilities;
 
-namespace Bit.Core.OrganizationFeatures.OrganizationApiKeys;
+namespace Bit.Core.LoginFeatures.PasswordlessLogin;
 
 public class VerifyAuthRequestCommand : IVerifyAuthRequestCommand
 {
